Full Oli4v3 demo and desktop version as main.py
The Oli4v3 demo is done, added LED indicator of process status, scene structuring and framework, individual prompt and stopword options for each scene and choosing between sitting and standing.

Also we were just doing call-response and not actual conversations (oops). Added conversation capabilities with history.

Also created a desktop demo of this version without robot functionality so you can run it on pc to work out the prompts, test latency of models and appropriateness of gestures. install pyttsx3 to make it work.