Edited main.py and changed setup
You can now install the environment as a venv in the same folder using install.py. run_all.bat now uses this venv to make installing and running easier. The GestureAPI script and model is now stored in the project folder. It automatically installs, which is needed since it is too large to commit to git